fix(resolver): it now yields the correct dataset for last+before args
Target the issue nodkz/graphql-compose-connection#7
In backward situations with cursors, the resolver uses count to compute
the "skip" value to forward to the findMany resolver. If the count value
is the global one, and not the one with the rawQuery with the '$lt'
statement, the value is incorrect. Hence the invalid dataset.